Fix padding of minuet-bold via line-height
The line-height of 0 essentially set the height to basically nothing therefore it had absolutely no padding and felt squashed, the text exceeded the bounds of the minuet.

With it being changed to 10 (and not normal) it's specifically enough to be padded and fit the text comfily, but not so much as to stretch the containers bounds.